Fix handling of event counters
The 'ack_requests' function is badly named because it has
no side-effect on the ring: it simply advances the host's private
'consumer' pointer and upcalls to the application.

The C macros have a 'final check for requests' concept where
the backend, which is about to block, writes the event pointer
to point to the next index it would like to read from. If a
frontend sees the event pointer with a low number, it means that
the backend is still working and will eventually see the new
data. If the frontend sees the event pointer appear in the middle
of the slots it has just written, it must signal the backend to
wake up.

Note this means that the backend should be written in a synchronous
style, with one single thread owning the ring, managing the pointers
and blocking. It is unsafe to let parallel threads access the ring.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>