RPC frontends: actually call the initialise function
If we forget to do this then all looks superficially ok except
the {req,rsp}_event counters are 0 rather than 1. This means that
the backend will believe it doesn't have to notify the frontend
and so the frontend blocks forever.

This was previously being masked by receiving extra wakeups from
Activations.wait.