Add Ring.Rpc.of_buf_no_init
It seems that the only way to get an `Rpc.sring` was to use `of_buf`,
which also initialises the ring. This is a problem if the other party
has already initialised it, as it overwrites their data.

In particular, this prevented Qubes disposable Linux VMs from working
when restored. They would populate the ring with RX pages before handing
it to the Mirage netback, which would then reinitialise the ring and
lose them all. It would then wait forever for Linux to provide it with
some pages to write to.

At least, this is my guess. The protocol does not appear to be
documented anywhere.